move UserController.SaveGame to SavedGameController.saveGame moredetail see desc
first create
SavedGameController.saveGame
move UserController.SaveGame to SavedGameController.saveGame moredetail see desc
