+++ chessxone-backend/api/routes/helpers/savedGames.js
@@ -1,0 +1,7 @@
+const router = require("express").Router();
+
+const savedGameController = require("../../controllers/savedGame")
+
+router.put('/:gameID/personal/:userID',savedGameController.saveGame)
+
+module.exports = router;
